requirements.txt
google-cloud-bigquery==1.24.0
google-cloud-pubsub==1.4.2
google-api-python-client==1.8.0